omap2+: Move nand and memory init out of s_init()
The contents of .bss section are cleared after
s_init(). This includes the flag "is_ddr_166M".
Even though, value assigned to this flag is used
immediately to configure SDRC, the flag is cleared
and re-assigned later.

It is not *really* necessary to initialize both
the NAND chip and the memory so early in the boot
process.

This patch delays bothe operations further down
the sequence along with other initializations.
This also ensures that init steps are executed
once.

Delayed initialization shouldn't impact anything
because memory initialization needs to be done
before loading the u-boot. This is still true.

Similar changes were necessary for AM3517EVM due
to cascading effect. But, it helps in maintaining
same flow through x-loader.

In addition:
  * Rename config_3430sdram_ddr() to sdrc_init()
    to use  prototype defined in sys_proto.h
  * Rename config_emif4_ddr() to emif_init() and
    add prototype definition in in sys_proto.h
  * Remove unused prototype do_sdrc_init().

Signed-off-by: Sanjeev Premi <premi@ti.com>